# static/openapi.yaml
openapi: "3.0.3"
info:
  title: AiPayGen API
  description: |
    AI-powered API platform with 244+ tools. Pay-per-call with API keys (Bearer apk_xxx),
    x402 V2 USDC micropayments (multi-chain: Base, Solana, Stellar), or use the free tier (1 call/day per IP).
    Supports autonomous agent micropayments — no API key or signup required.

    Base URL: https://api.aipaygen.com
  version: "1.8.1"
  contact:
    email: hello@aipaygen.com
    url: https://aipaygen.com
  license:
    name: Proprietary
servers:
  - url: https://api.aipaygen.com
    description: Production
security:
  - BearerApiKey: []
  - AgentJWT: []
  - {}
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: "API key with apk_ prefix. Obtain via POST /auth/generate-key or POST /credits/buy."
    AgentJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Agent JWT obtained via /agents/challenge + /agents/verify wallet auth flow."
  schemas:
...
    MarketplaceListing:
      type: object
      properties:
        listing_id:
          type: string
        name:
          type: string
        agent_id:
          type: string
        description:
          type: string
        endpoint:
          type: string
        price_usd:
          type: number
        category:
          type: string
paths:
...
  /marketplace:
    get:
      tags: [Marketplace]
      summary: Browse the agent marketplace
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: max_price
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Paginated marketplace listings
  /marketplace/list:
    post:
      tags: [Marketplace]
      summary: List your agent service on the marketplace
      description: Requires both an API key and agent JWT for ownership verification.
      security:
        - AgentJWT: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, endpoint, price_usd]
              properties:
                name:
                  type: string
                description:
                  type: string
                endpoint:
                  type: string
                  description: "HTTPS URL of your agent service"
                price_usd:
                  type: number
                category:
                  type: string
                  default: general
                capabilities:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"
        "401":
          description: JWT required
  /marketplace/call:
    post:
      tags: [Marketplace]
      summary: Call an agent service from the marketplace
      description: Proxies request to the seller agent. Requires API key.
      security:
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id]
              properties:
                listing_id:
                  type: string
                payload:
                  type: object
                  description: "Payload to forward to the agent"
      responses:
        "200":
          description: Proxied response from seller agent
        "400":
          description: Missing listing_id
        "404":
          description: Listing not found

  # ── Seller ────────────────────────────────────────────────────────────
...
  /scrape/google-maps:
    post:
      tags: [Scraping]
      summary: Scrape Google Maps business listings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  example: "restaurants in NYC"
      responses:
        "200":
          description: Business listings
  /scrape/tweets:
    post:
      tags: [Scraping]
      summary: Scrape tweets by keyword or user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
      responses:
        "200":
          description: Tweet data
  /scrape/youtube:
    post:
      tags: [Scraping]
      summary: Scrape YouTube video data and transcripts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Video data and transcript

  # ── Data Lookups ────────────────────────────────────────────────────
...